% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_vis.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gg_miss_upset}
\alias{md.pattern}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mice}{\code{\link[mice]{md.pattern}}}

  \item{naniar}{\code{\link[naniar]{gg_miss_upset}}}
}}

