% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi-package.R
\docType{package}
\name{smdi-package}
\alias{smdi}
\alias{smdi-package}
\title{smdi: Perform Structural Missing Data Investigations}
\description{
An easy to use implementation of routine structural missing data diagnostics with functions to visualize the proportions of missing observations, investigate missing data patterns and conduct various empirical missing data diagnostic tests. Reference: Weberpals J, Raman SR, Shaw PA, Lee H, Hammill BG, Toh S, Connolly JG, Dandreo KJ, Tian F, Liu W, Li J, Hernández-Muñoz JJ, Glynn RJ, Desai RJ. smdi: an R package to perform structural missing data investigations on partially observed confounders in real-world evidence studies. JAMIA Open. 2024 Jan 31;7(1):ooae008. \doi{10.1093/jamiaopen/ooae008}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://janickweberpals.gitlab-pages.partners.org/smdi/}
  \item Report bugs at \url{https://gitlab-scm.partners.org/janickweberpals/smdi/-/issues}
}

}
\author{
\strong{Maintainer}: Janick Weberpals \email{janick.developer@gmail.com} (\href{https://orcid.org/0000-0003-0404-7394}{ORCID}) [copyright holder]

}
\keyword{internal}
