% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_asmd.R
\name{smdi_asmd}
\alias{smdi_asmd}
\title{Compute mean/median absolute standardized mean differences}
\usage{
smdi_asmd(
  data = NULL,
  covar = NULL,
  na_strategy = c("retain", "drop"),
  median = TRUE,
  ...
)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with variable/column name(s) to investigate}

\item{na_strategy}{"retain" or "drop" covariates after creating misisng indicator variables; more info see ?smdi::smdi_na_indicator()}

\item{median}{logical if the median (recommended default) or mean of all absolute standardized mean differences should be computed}

\item{...}{further arguments}
}
\value{
returns mean/median absolute standardized mean differences
}
\description{
This function takes a dataframe and covariates which are only partially observed/missing.
The dataframe should consist of the exposure variable, the outcome variable(s), the partially observed covariates
and all other fully observed covariates which are deemed important for the final modeling
or which could be considered as auxiliary variables. If no partially observed covariates are provided,
the function automatically looks for all variables/columns with NA.

Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\examples{
\dontrun{
library(smdi)
library(dplyr)

df <- smdi_data \%>\%
select(-id) \%>\%
mutate(across(ends_with("cat"), as.factor))
df \%>\%
smdi_asmd(covar = c("age", "gender", "bmi"))
}
}
