% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_check_covar.R
\name{smdi_check_covar}
\alias{smdi_check_covar}
\title{This is a util function to help check input data and covariates provided}
\usage{
smdi_check_covar(data = NULL, covar = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation}
}
\value{
returns the covariate vector for subsequent tasks or warnings/errors
}
\description{
This is a util function to help check input data and covariates provided
}
