% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_hotelling.R
\name{smdi_hotelling}
\alias{smdi_hotelling}
\title{Computes hotelling's multivariate t-test}
\usage{
smdi_hotelling(data = NULL, covar = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with partially observed variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation and all remaining covariates will be used as predictors}

\item{median}{logical if the median (= TRUE; recommended default) or mean of all absolute standardized mean differences (asmd) should be computed}

\item{includeNA}{logical, should missingness of other partially observed covariates be explicitly modeled (default is FALSE)}

\item{...}{further arguments}
}
\value{
returns an asmd object with mean/median absolute standardized mean differences
}
\description{
Hotelling's multivariate t-test, which examines variable
differences conditional on having an observed covariate value or not.
As the power of statistical hypothesis tests can be influenced by
sample size, the combined investigation along with smdi_asmd() is highly recommended.
}
\details{
Wrapper of
#' Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\examples{
\dontrun{
library(smdi)
library(dplyr)

}
}
\references{
Hotelling H. The Generalization of Student’s Ratio. Ann Math Stat. 1931;2(3):360-378. doi:10.1214/aoms/1177732979
}
