% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi-package.R
\docType{package}
\name{smdi-package}
\alias{smdi}
\alias{smdi-package}
\title{smdi: Perform Structural Missing Data Investigations}
\description{
This package aims to help implementing routine structural missing data diagnostics. It provides functions to: visualize proportions of missing observations, investigate missing data patterns and perform various empirical missing data diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://janickweberpals.gitlab-pages.partners.org/smdi}
  \item Report bugs at \url{https://gitlab-scm.partners.org/janickweberpals/smdi/-/issues}
}

}
\author{
\strong{Maintainer}: Janick Weberpals \email{jweberpals@bwh.harvard.edu} (\href{https://orcid.org/0000-0003-0404-7394}{ORCID}) [copyright holder]

}
\keyword{internal}
