% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_data.R
\docType{data}
\name{smdi_data}
\alias{smdi_data}
\title{smdi exemplary lung cancer dataset}
\format{

}
\source{
https://janickweberpals.gitlab-pages.partners.org/smdi/articles/data_generation.html
}
\usage{
smdi_data
}
\description{
Example dataset with only complete cases.
}
\keyword{datasets}
