% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_data_complete.R
\docType{data}
\name{smdi_data_complete}
\alias{smdi_data_complete}
\title{smdi exemplary lung cancer dataset (with complete data)}
\format{

}
\source{
https://janickweberpals.gitlab-pages.partners.org/smdi/articles/data_generation.html
}
\usage{
smdi_data_complete
}
\description{
Example dataset with complete cases.
}
\keyword{datasets}
