% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_diagnose.R
\name{smdi_diagnose}
\alias{smdi_diagnose}
\title{Computes three group missing data summary diagnostics}
\usage{
smdi_diagnose(
  data = NULL,
  covar = NULL,
  model = c("logistic", "linear", "cox"),
  form_lhs = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with partially observed variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation and all remaining covariates will be used as predictors}

\item{model}{character describing which outcome model to fit to assess the association between covar missingness indicator and outcome. Currently supported are models of type logistic, linear and cox (see smdi_outcome)}

\item{form_lhs}{string specifying the left-hand side of the outcome formula (see smdi_outcome)}

\item{...}{further arguments to pass on to smdi_asmd, smdi_hotelling, smdi_little, smdi_rf or smdi_outcome}

\item{exponentiated}{logical, should results of the fitted outcome model be exponentiated (see smdi_outcome)}
}
\value{
smdi object with details of all wrapped functions. print() returns convenient summary table
}
\description{
This function

#' #' Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\details{
The function automatically fits a crude and adjusted outcome model. The currently supported models are logistic (glm), linear (lm) and cox (survival).
For adjusted models, the function uses all available covariates found in the dataset specified with the <data> parameter. If covariates should not
be include in the outcome model, these covariates should be dropped beforehand (as with all other functions in the smdi package).

The left-hand side of the formula (<form_lhs>) needs to specify the outcome in one of the following ways:
- glm (binary): character of column name with binary outcome, e.g. "MACE"
- lm (continuous): character of column name with binary outcome, e.g. "WEIGHT_LOSS"
- cox (time-to-event): LHS specifying time-to-event outcome, e.g. "Surv(TIME, STATUS)"
}
\examples{
library(smdi)

smdi_outcome(
 data = smdi_data,
 model = "cox",
 form_lhs = "Surv(eventtime, status)"
)


}
\references{
...
}
\seealso{
[smdi_asmd()]
[smdi_hotteling()]
[smdi_little()]
[smdi_rf()]
[smdi_outcome()]
}
