% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_hotelling.R
\name{smdi_hotelling}
\alias{smdi_hotelling}
\title{Computes hotelling's multivariate t-test}
\usage{
smdi_hotelling(data = NULL, covar = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with partially observed variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation and all remaining covariates will be used as predictors}
}
\value{
returns a hotelling object with statistics on hotellings test by covariate. That is, for each <covar>, the following outputs are provided:

- stats: hotelling test statistics (for more information see \code{\link{hotelling.test}})
}
\description{
Hotelling's multivariate t-test, which examines variable
differences conditional on having an observed covariate value or not.
As the power of statistical hypothesis tests can be influenced by
sample size, the combined investigation along with smdi_asmd() is highly recommended.

Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\details{
...
}
\examples{
\dontrun{
library(smdi)

smdi_hotteling(data = smdi_data)

}
}
\references{
Hotelling H. The Generalization of Student’s Ratio. Ann Math Stat. 1931;2(3):360-378. doi:10.1214/aoms/1177732979
}
\seealso{
\code{\link{hotelling.test}}
}
