% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_little.R
\name{smdi_little}
\alias{smdi_little}
\title{Computes Little's test}
\usage{
smdi_little(data = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with partially observed variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation and all remaining covariates will be used as predictors}
}
\value{
returns a hotelling object with statistics on hotellings test by covariate. S3 method print returns a summarized dataframe with hypothesis test p-values.
}
\description{
Little’s chi-squared test takes into account possible patterns of missingness across all variables in the dataset.
Rejection of the null hypothesis of this test would provide sufficient evidence to indicate that the data are (globally) not MCAR.
Please note that compared to smdi_hotelling, this function tests for MCAR globally across all missing covariates.

#' #' Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\details{
Wrapper of mcar_test (naniar package)
}
\examples{
\dontrun{
library(smdi)
library(dplyr)

smdi_data \%>\%
 smdi_little()

}
}
\references{
Little RJA. A Test of Missing Completely at Random for Multivariate Data with Missing Values.
	J Am Stat Assoc. 1988;83(404):1198-1202. doi:10.1080/01621459.1988.10478722
}
