% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_narfcs.R
\name{smdi_narfcs}
\alias{smdi_narfcs}
\title{Wrapper of NARFCS enhanced forked mice version: https://raw.githack.com/moreno-betancur/NARFCS/master/Vignette.html}
\usage{
smdi_narfcs(data = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}
}
\value{
returns a table with covariate name, amount missing observations and proportion missing.
}
\description{
This function takes a dataframe and returns all columns which are partially observed with
the amount of observations missing and corresponding proportion assuming a one-row-per-patient
dataframe.

Important: avoid to include variables like ID variables, ZIP codes, dates, etc.
}
\examples{
\dontrun{

df \%>\%
smdi_narfcs()
}
}
