% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_rf.R
\name{smdi_rf}
\alias{smdi_rf}
\title{Computes random forest-based AUC}
\usage{
smdi_rf(
  data = NULL,
  covar = NULL,
  train_test_ratio = c(0.7, 0.3),
  set_seed = 42,
  ntree = 1000
)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with partially observed variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation and all remaining covariates will be used as predictors}

\item{train_test_ratio}{numeric vector to indicate the test/train split ratio, e.g. c(.7, .3) which is the default}

\item{set_seed}{seed for reproducibility, defaults to 42}

\item{ntree}{integer, number of trees (defaults to 1000 trees)}
}
\value{
rf object: list that contains the ROC AUC value and corresponding variable importance in training dataset (latter as ggplot object)
}
\description{
The function trains and fits a random forest model to assess the ability to predict missingness for
the specified covariate(s). If missing indicator can be predicted as a function of observed covariates,
MAR may be a likely scenario and would imply that imputation may be feasible.

Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\details{
The random forest utilizes the randomForest engine.
}
\examples{
library(smdi)

smdi_rf(data = smdi_data)
}
\references{
Weberpals J,  Sondhi A, Jiang C, Yerram P, Taylor MD,  Samant M, Cherng ST. A Systematic Approach Towards Missing Lab Data in Electronic Health Records: A Case Study in Non-Small Cell Lung Cancer and Multiple Myeloma. 37th International Conference on Pharmacoepidemiology & Therapeutic Risk Management 2021. Pharmacoepidemiol Drug Saf 2021; 30:36-36.
https://onlinelibrary.wiley.com/doi/10.1002/pds.5305
}
\seealso{
\code{\link{randomForest}}
}
