% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_rf.R
\name{smdi_rf}
\alias{smdi_rf}
\title{Computes random forest-based AUC}
\usage{
smdi_rf(
  data = NULL,
  covar = NULL,
  tune = FALSE,
  train_test_ratio = c(0.7, 0.3),
  set_seed = 42,
  ...
)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with partially observed variable/column name(s) to investigate. If NULL, the function automatically includes all columns with at least one missing observation and all remaining covariates will be used as predictors}

\item{tune}{logical if random forest model for AUC diagnostic should be tuned and cross-validated (CAVE: very time consuming, default is FALSE)}

\item{train_test_ratio}{numeric vector to indicate the test/train split ratio, e.g. c(.7, .3) which is the default}

\item{set_seed}{seed for reproducibility, defaults to 42}

\item{...}{additional arguments}
}
\value{
...
}
\description{
The function trains and fits a random forest model to assess the ability to predict missingness for
the specified covariate(s). If missing indicator can be predicted as a function of observed covariates,
MAR may be a likely scenario and would imply that imputation may be feasible.

#' #' Important: don't include variables like ID variables, ZIP codes, dates, etc.
}
\details{
...
}
\examples{
\dontrun{
library(smdi)
library(dplyr)

smdi_data \%>\%
 smdi_little()

}
}
\references{
Weberpals J, Sondhi A., et al.
	[Pharmacoepidemiol Drug Saf, Vol. 30, pp. 36-36, 2021](https://onlinelibrary.wiley.com/doi/10.1002/pds.5305),
	presented as an oral presentation at the International Conference on
	Pharmacoepidemiology and Risk Management (2021)
}
