% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_summarize.R
\name{smdi_summarize}
\alias{smdi_summarize}
\title{Util helper to automatically extract all covariates with partially observed covariates}
\usage{
smdi_summarize(data = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}
}
\value{
returns a table with covariate name, amount missing observations and proportion missing.
}
\description{
This function takes a dataframe and automatically returns all columns which are partially observed with
the amount of observations missing and corresponding proportion assuming a one-row-per-patient
dataframe. This is an important utility function for many other functions in this package

Important: avoid to include variables like ID variables, ZIP codes, dates, etc.
}
\examples{
\dontrun{

df \%>\%
smdi_summarize()
}
}
