% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdi_vis.R
\name{smdi_vis}
\alias{smdi_vis}
\title{Quick and elegant visualization of partially observed/missing variables}
\usage{
smdi_vis(data = NULL, covar = NULL, top_n_covar = NULL, strata = NULL)
}
\arguments{
\item{data}{dataframe or tibble object with partially observed/missing variables}

\item{covar}{character covariate or covariate vector with variable/column name(s) to investigate}

\item{top_n_covar}{integer, display top n missing covariates}

\item{strata}{character name of variable/column by which results should be stratified}
}
\value{
returns ggplot2 graph displaying selected or chosen variables by percent missing
}
\description{
This function takes a dataframe and outputs a nicely formatted
ggplot2 vertical barchart plot that visualizes the proportion
missing for a given variable (vector) or the top n missing
variables. Results can also be stratified by another variable.
The function assumes a one-row-per-patient dataframe.
}
\examples{
\dontrun{
library(smdi)
library(dplyr)

smdi_data \%>\%
smdi_vis()

}
}
